use guard for continuous testing
Watch mode of mocha is unreliable, sometimes takes a couple of seconds
to notice a change.

I tried guard-mocha-node
    https://github.com/neerolyte/guard-mocha-node
but could not get it to work. So guard-shell.

However, guard-shell hated mocha's console funkiness and scrolled past the failure when I used `--clear`. This prompted me to use the unfunk reporter. Coincidentally this reporter also fixes mocha's color madness.

The only pity is that the unfunk reporter always shows all specs.